+++ builds/compromise.js
@@ -1,4 +1,4 @@
-/* compromise 13.3.0 MIT */
+/* compromise 13.3.1 MIT */
 (function (global, factory) {
   typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
   typeof define === 'function' && define.amd ? define(factory) :
@@ -3132,8 +3132,10 @@
 
     if (!text || typeof text !== 'string' || hasSomething.test(text) === false) {
       return sentences;
-    } // Start somewhere:
+    } // cleanup unicode-spaces
 
+
+    text = text.replace('\xa0', ' '); // Start somewhere:
 
     var splits = naiive_split(text); // Filter-out the crap ones
 
@@ -3429,7 +3431,7 @@
 
   var fromJSON_1 = fromJSON;
 
-  var _version = '13.3.0';
+  var _version = '13.3.1';
 
   var _data = {
     "Comparative": "true¦better",