+++ plugins/dates/src/02-getDate/index.js
@@ -2,12 +2,12 @@ const spacetime = require('spacetime')
 const normalize = require('./00-normalize')
 const parseRanges = require('./01-parse-range')
 
-const getDate = function(doc, context) {
+const getDate = function (doc, context) {
   // validate context a bit
   context = context || {}
 
   context.timezone = context.timezone || 'ETC/UTC'
-  context.today = spacetime(context.today, context.timezone)
+  context.today = spacetime(context.today || null, context.timezone)
   //turn 'five' into 5..
   doc = normalize(doc)
   //interpret 'between [A] and [B]'...