+++ plugins/dates/src/03-parseDate/steps/03-explicit.js
@@ -1,28 +1,30 @@
 const { Unit, Day, CalendarDate } = require('../_units')
 
 const knownWord = {
-  today: context => new Day(context.today, null, context),
-  yesterday: context => {
+  today: (context) => new Day(context.today, null, context),
+  yesterday: (context) => {
     new Day(context.today.minus(1, 'day'), null, context)
   },
-  tomorrow: context => {
+  tomorrow: (context) => {
     new Day(context.today.plus(1, 'day'), null, context)
   },
 }
 
 // parse things like 'june 5th 2019'
-const parseExplicit = function(doc, context) {
+// most of this is done in spacetime
+const parseExplicit = function (doc, context) {
+  let impliedYear = context.today.year()
   // 'fifth of june'
-  let m = doc.match('[<date>#Value] of [<month>#Month]')
+  let m = doc.match('[<date>#Value] of [<month>#Month] [<year>#Year?]')
+  // 'june the fifth'
   if (!m.found) {
-    // 'june the fifth'
-    m = doc.match('[<month>#Month] the [<date>#Value]')
+    m = doc.match('[<month>#Month] the [<date>#Value] [<year>#Year?]')
   }
   if (m.found) {
     let obj = {
       month: m.groups('month').text(),
       date: m.groups('date').text(),
-      year: context.today.year(),
+      year: m.groups('year').text() || impliedYear,
     }
     let d = new CalendarDate(obj, null, context)
     if (d.d.isValid() === true) {
@@ -48,10 +50,8 @@ const parseExplicit = function(doc, context) {
     let d = knownWord[str](context)
     return d
   }
-
   // punt it to spacetime, for the heavy-lifting
   let d = new Unit(str, null, context)
-  // console.log(context.d.format('nice-year'))
   // did we find a date?
   if (d.d.isValid() === false) {
     return null