+++ plugins/dates/src/index.js
@@ -2,8 +2,9 @@ const tagger = require('./01-tag')
 const tags = require('./data/_tags')
 const words = require('./data/words')
 const methods = require('./methods')
+const spacetime = require('spacetime')
 
-const addMethods = function(Doc, world) {
+const addMethods = function (Doc, world) {
   // our new tags
   world.addTags(tags)
   // add info for the date plugin
@@ -15,13 +16,15 @@ const addMethods = function(Doc, world) {
   class Dates extends Doc {
     constructor(list, from, w) {
       super(list, from, w)
-      this.context = {}
+      this.context = {
+        casual_duration: { weeks: 2 },
+      }
     }
   }
   //add-in methods
   Object.assign(Dates.prototype, methods)
 
-  Doc.prototype.dates = function(n) {
+  Doc.prototype.dates = function (n) {
     let context = {}
     if (n && typeof n === 'object') {
       context = n
@@ -36,6 +39,9 @@ const addMethods = function(Doc, world) {
       dates = dates.get(n)
     }
     let d = new Dates(dates.list, this, this.world)
+    if (context.today) {
+      context.today = spacetime(context.today, context.timezone)
+    }
     d.context = context
     return d
   }